move StatisticSeries to database crate
